Seperated out webEndpoint tests
